+++ src/main/java/com/deeerui/money/config/TaxConfiguration.java
@@ -2,15 +2,17 @@ package com.deeerui.money.config;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
+import java.math.BigDecimal;
+
 @ConfigurationProperties("money.tax")
 public class TaxConfiguration {
-    private double rate;
+    private BigDecimal rate;
 
-    public double getRate() {
+    public BigDecimal getRate() {
         return rate;
     }
 
-    public void setRate(double rate) {
+    public void setRate(BigDecimal rate) {
         this.rate = rate;
     }
 }
